refactor: introduce project for user commands abstractions
