+++ Source/Application/Kysect.Shreks.Application.Commands/Kysect.Shreks.Application.Commands.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFramework>net6.0</TargetFramework>
@@ -17,10 +17,7 @@
     <ItemGroup>
       <ProjectReference Include="..\..\Domain\Kysect.Shreks.Common\Kysect.Shreks.Common.csproj" />
       <ProjectReference Include="..\Kysect.Shreks.Application.Abstractions\Kysect.Shreks.Application.Abstractions.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Submissions\" />
+      <ProjectReference Include="..\Kysect.Shreks.Application.UserCommands.Abstractions\Kysect.Shreks.Application.UserCommands.Abstractions.csproj" />
     </ItemGroup>
 
 </Project>
