+++ Source/Application/Kysect.Shreks.Application.Commands/Processors/IShreksCommandVisitor.cs
@@ -1,5 +1,5 @@
-using Kysect.Shreks.Application.Commands.Commands;
 using Kysect.Shreks.Application.Commands.Result;
+using Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
 
 namespace Kysect.Shreks.Application.Commands.Processors;
 
@@ -13,4 +13,22 @@ public interface IShreksCommandVisitor<TResult> where TResult : IShreksCommandRe
     Task<TResult> VisitAsync(DeactivateCommand command);
     Task<TResult> VisitAsync(CreateSubmissionCommand command);
     Task<TResult> VisitAsync(DeleteCommand command);
+}
+
+public static class ShreksCommandVisitorExtensions
+{
+    public static Task<TResult> Visit<TResult>(this IShreksCommandVisitor<TResult> visitor, IShreksCommand command) where TResult : IShreksCommandResult
+    {
+        return command switch
+        {
+            ActivateCommand activateCommand => visitor.Visit(activateCommand),
+            CreateSubmissionCommand createSubmissionCommand => visitor.Visit(createSubmissionCommand),
+            DeactivateCommand deactivateCommand => visitor.Visit(deactivateCommand),
+            DeleteCommand deleteCommand => visitor.Visit(deleteCommand),
+            HelpCommand helpCommand => visitor.VisitAsync(helpCommand),
+            RateCommand rateCommand => visitor.VisitAsync(rateCommand),
+            UpdateCommand updateCommand => visitor.Visit(updateCommand),
+            _ => throw new ArgumentOutOfRangeException(nameof(command)),
+        };
+    }
 }
