+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/ActivateCommand.cs
@@ -1,0 +1,8 @@
+using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/activate")]
+public class ActivateCommand : IShreksCommand
+{
+}
