+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/CreateSubmissionCommand.cs
@@ -1,0 +1,8 @@
+﻿using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/create-submission")]
+public class CreateSubmissionCommand : IShreksCommand
+{
+}
