+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/DeactivateCommand.cs
@@ -1,0 +1,8 @@
+using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/deactivate")]
+public class DeactivateCommand : IShreksCommand
+{
+}
