+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/DeleteCommand.cs
@@ -1,0 +1,9 @@
+using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/delete")]
+public class DeleteCommand : IShreksCommand
+{
+
+}
