+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/HelpCommand.cs
@@ -1,0 +1,8 @@
+﻿using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/help")]
+public class HelpCommand : IShreksCommand
+{
+}
