+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/IShreksCommand.cs
@@ -1,0 +1,5 @@
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+public interface IShreksCommand
+{
+}
