+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/RateCommand.cs
@@ -1,0 +1,19 @@
+using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/rate", aliases: new []{"/assess"})]
+public class RateCommand : IShreksCommand
+{
+    public RateCommand(double ratingPercent, double? extraPoints)
+    {
+        RatingPercent = ratingPercent;
+        ExtraPoints = extraPoints;
+    }
+
+    [Value(0, Required = true, MetaName = "RatingPercent")]
+    public double RatingPercent { get; }
+    
+    [Value(1, Required = false, Default = 0.0, MetaName = "ExtraPoints")]
+    public double? ExtraPoints { get; }
+}
