+++ Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/UpdateCommand.cs
@@ -1,0 +1,27 @@
+using CommandLine;
+
+namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+
+[Verb("/update")]
+public class UpdateCommand : IShreksCommand
+{
+    public UpdateCommand(int submissionCode, double? ratingPercent, double? extraPoints, string? dateStr)
+    {
+        SubmissionCode = submissionCode;
+        RatingPercent = ratingPercent;
+        ExtraPoints = extraPoints;
+        DateStr = dateStr;
+    }
+
+    [Value(0, Required = true, MetaName = "SubmissionCode")]
+    public int SubmissionCode { get; }
+    
+    [Option(shortName:'r', longName: "rating", Group = "update",  Required = false)]
+    public double? RatingPercent { get; }
+    
+    [Option(shortName:'e', longName:"extra", Group = "update", Required = false)]
+    public double? ExtraPoints { get; }
+    
+    [Option(shortName:'d', longName:"date", Group = "update", Required = false)]
+    public string? DateStr { get; }
+}
