+++ Source/Kysect.Shreks.sln
@@ -59,9 +59,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.ApiClientGene
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.DataImport", "Playground\Kysect.Shreks.DataImport\Kysect.Shreks.DataImport.csproj", "{601BCCAB-8895-43CB-BC46-EE2C20F4031F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Application.GithubWorkflow", "Application\Kysect.Shreks.Application.GithubWorkflow\Kysect.Shreks.Application.GithubWorkflow.csproj", "{DF0C2310-8F92-404C-885E-D495B3F334EB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Application.GithubWorkflow", "Application\Kysect.Shreks.Application.GithubWorkflow\Kysect.Shreks.Application.GithubWorkflow.csproj", "{DF0C2310-8F92-404C-885E-D495B3F334EB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Application.GithubWorkflow.Abstractions", "Application\Kysect.Shreks.Application.GithubWorkflow.Abstractions\Kysect.Shreks.Application.GithubWorkflow.Abstractions.csproj", "{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Application.GithubWorkflow.Abstractions", "Application\Kysect.Shreks.Application.GithubWorkflow.Abstractions\Kysect.Shreks.Application.GithubWorkflow.Abstractions.csproj", "{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Application.UserCommands.Abstractions", "Application\Kysect.Shreks.Application.UserCommands.Abstractions\Kysect.Shreks.Application.UserCommands.Abstractions.csproj", "{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -161,6 +163,10 @@ Global
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -190,6 +196,7 @@ Global
 		{601BCCAB-8895-43CB-BC46-EE2C20F4031F} = {DC866D57-D41E-455D-BD7D-85F1DD575823}
 		{DF0C2310-8F92-404C-885E-D495B3F334EB} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
+		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CD5C67DB-70F8-4C47-A3F4-7C94BECE1657}
